Port eval functions to Rust and remove C stubs
