--- rust_evalfunc_stubs/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "rust_evalfunc_stubs"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-libc = "0.2"
