+++ src/Make_ami.mak
@@ -93,8 +93,7 @@ SRC += \
 	edit.c \
 	eval.c \
 	evalbuffer.c \
-	evalfunc.c \
-	evalvars.c \
+        evalvars.c \
         evalwindow.c \
 	ex_cmds2.c \
 	ex_docmd.c \
