+++ src/Makefile
@@ -1391,8 +1391,6 @@ RUST_PYTHON_DIR = ../rust_python
 RUST_PYTHON_LIB = $(RUST_PYTHON_DIR)/target/release/librust_python.a
 RUST_EVALFUNC_DIR = ../rust_evalfunc
 RUST_EVALFUNC_LIB = $(RUST_EVALFUNC_DIR)/target/release/librust_evalfunc.a
-RUST_EVALFUNC_STUBS_DIR = ../rust_evalfunc_stubs
-RUST_EVALFUNC_STUBS_LIB = $(RUST_EVALFUNC_STUBS_DIR)/target/release/librust_evalfunc_stubs.a
 # Rust eval crate exposing eval_expr_rs used by C side
 RUST_EVAL_DIR = ../rust_eval
 RUST_EVAL_LIB = $(RUST_EVAL_DIR)/target/release/librust_eval.a
@@ -1473,7 +1471,6 @@ ALL_LIBS = \
            $(RUST_CHANNEL_LIB) \
            $(RUST_GUI_LIB) \
            $(RUST_EVALFUNC_LIB) \
-           $(RUST_EVALFUNC_STUBS_LIB) \
            $(RUST_EVAL_LIB) \
            $(RUST_VIM9_LIB) \
            $(RUST_EXCMD_LIB) \
@@ -1575,10 +1572,7 @@ $(RUST_GUI_LIB):
 	cd $(RUST_GUI_DIR) && CARGO_TARGET_DIR=target cargo build --release
 
 $(RUST_EVALFUNC_LIB):
-       cd $(RUST_EVALFUNC_DIR) && CARGO_TARGET_DIR=target cargo build --release
-
-$(RUST_EVALFUNC_STUBS_LIB):
-       cd $(RUST_EVALFUNC_STUBS_DIR) && CARGO_TARGET_DIR=target cargo build --release
+        cd $(RUST_EVALFUNC_DIR) && CARGO_TARGET_DIR=target cargo build --release
 
 $(RUST_EVAL_LIB):
 	cd $(RUST_EVAL_DIR) && CARGO_TARGET_DIR=target cargo build --release
@@ -1794,7 +1788,6 @@ BASIC_SRC = \
         edit.c \
         eval.c \
         evalbuffer.c \
-        evalfunc.c \
         evalvars.c \
         evalwindow.c \
         ex_eval.c \
@@ -2059,8 +2052,7 @@ PRO_AUTO = \
 	edit.pro \
 	eval.pro \
 	evalbuffer.pro \
-	evalfunc.pro \
-	evalvars.pro \
+   evalvars.pro \
 	evalwindow.pro \
 	ex_cmds.pro \
 	ex_cmds2.pro \
@@ -2263,7 +2255,7 @@ CCC = $(CCC_NF) $(ALL_CFLAGS)
 
 # Link the target for normal use or debugging.
 # A shell script is used to try linking without unnecessary libraries.
-$(VIMTARGET): auto/config.mk $(OBJ) objects/version.o $(RUST_MEM_LIB) $(RUST_REGEX_LIB) $(SRC_RUST_REGEX_LIB) $(RUST_BUFFER_LIB) $(RUST_BUFWRITE_LIB) $(RUST_CHANNEL_LIB) $(RUST_GUI_LIB) $(RUST_EVALFUNC_LIB) $(RUST_EVALFUNC_STUBS_LIB) $(RUST_EVAL_LIB) $(RUST_VIM9_LIB) $(RUST_EXCMD_LIB) $(RUST_CMDHIST_LIB) $(RUST_ARABIC_LIB) $(RUST_BEVAL_LIB) $(RUST_TAG_LIB) $(RUST_ALLOC_LIB) $(RUST_BLOB_LIB) $(RUST_SHA256_LIB)
+$(VIMTARGET): auto/config.mk $(OBJ) objects/version.o $(RUST_MEM_LIB) $(RUST_REGEX_LIB) $(SRC_RUST_REGEX_LIB) $(RUST_BUFFER_LIB) $(RUST_BUFWRITE_LIB) $(RUST_CHANNEL_LIB) $(RUST_GUI_LIB) $(RUST_EVALFUNC_LIB) $(RUST_EVAL_LIB) $(RUST_VIM9_LIB) $(RUST_EXCMD_LIB) $(RUST_CMDHIST_LIB) $(RUST_ARABIC_LIB) $(RUST_BEVAL_LIB) $(RUST_TAG_LIB) $(RUST_ALLOC_LIB) $(RUST_BLOB_LIB) $(RUST_SHA256_LIB)
 	@$(BUILD_DATE_MSG)
     @LINK="$(PURIFY) $(SHRPENV) $(CClink) $(ALL_LIB_DIRS) $(LDFLAGS) \
 		-o $(VIMTARGET) objects/*.o $(ALL_LIBS)" \
@@ -3310,11 +3302,8 @@ objects/eval.o: eval.c
 objects/evalbuffer.o: evalbuffer.c
 	$(CCC) -o $@ evalbuffer.c
 
-objects/evalfunc.o: evalfunc.c
-	$(CCC) -o $@ evalfunc.c
-
 objects/evalvars.o: evalvars.c
-	$(CCC) -o $@ evalvars.c
+        $(CCC) -o $@ evalvars.c
 
 objects/evalwindow.o: evalwindow.c
 	$(CCC) -o $@ evalwindow.c
@@ -3814,14 +3803,9 @@ objects/evalbuffer.o: evalbuffer.c vim.h protodef.h auto/config.h feature.h \
  beval.h proto/gui_beval.pro structs.h regexp.h gui.h \
  libvterm/include/vterm.h libvterm/include/vterm_keycodes.h alloc.h \
  ex_cmds.h spell.h proto.h globals.h errors.h
-objects/evalfunc.o: evalfunc.c vim.h protodef.h auto/config.h feature.h os_unix.h \
- auto/osdef.h ascii.h keymap.h termdefs.h macros.h option.h beval.h \
- proto/gui_beval.pro structs.h regexp.h gui.h libvterm/include/vterm.h \
- libvterm/include/vterm_keycodes.h alloc.h ex_cmds.h spell.h proto.h \
- globals.h errors.h version.h
 objects/evalvars.o: evalvars.c vim.h protodef.h auto/config.h feature.h os_unix.h \
- auto/osdef.h ascii.h keymap.h termdefs.h macros.h option.h beval.h \
- proto/gui_beval.pro structs.h regexp.h gui.h libvterm/include/vterm.h \
+        auto/osdef.h ascii.h keymap.h termdefs.h macros.h option.h beval.h \
+        proto/gui_beval.pro structs.h regexp.h gui.h libvterm/include/vterm.h \
  libvterm/include/vterm_keycodes.h alloc.h ex_cmds.h spell.h proto.h \
  globals.h errors.h version.h
 objects/evalwindow.o: evalwindow.c vim.h protodef.h auto/config.h feature.h \
